requirements.txt
streamlit
pandas
plotly
numpy
vnstock
PyPortfolioOpt
scipy
statsmodels
google-generativeai

# Database
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# Additional utilities
tqdm>=4.65.0
